feat: vision analyzer improvements — resize screenshots, better temperature, expanded error detection, deduplication
- Reduce screenshot size to 1280px width before sending to Vision AI (cheaper, faster, readable)
- Bump temperature from 0.3 to 0.5 for better visual judgment
- Expand error detection to catch blank, empty, frozen, broken, missing, spinner, stuck, etc.
- Add deduplication of repeated guidance events via analysis hashing
- Fix confidence type mismatch in engine.rs

+++ src/observation/src/engine.rs
@@ -407,7 +407,7 @@ impl ObservationEngine {
                                     .unwrap_or_default();
                                 let focused_app = data_obj.get("focused_app").and_then(|v| v.as_str()).map(|s| s.to_string());
                                 let user_activity = data_obj.get("user_activity").and_then(|v| v.as_str()).map(|s| s.to_string());
-                                let confidence = data_obj.get("confidence").and_then(|v| v.as_f64()).unwrap_or(0.0) as f32;
+                                let confidence = data_obj.get("confidence").and_then(|v| v.as_f64()).unwrap_or(0.0);
                                 let result = crate::vision_analyzer::VisionAnalysisResult {
                                     timestamp: chrono::Utc::now(),
                                     focused_app,
